net: Move run_vms to shared lib
To be used in new tests setup across different network test suites.
