+++ tests/network/localnet/ipam/conftest.py
@@ -7,6 +7,7 @@ from ocp_resources.namespace import Namespace
 import tests.network.libs.nodenetworkconfigurationpolicy as libnncp
 from libs.net import netattachdef as libnad
 from libs.net.ip import random_ipv4_address
+from libs.vm.oper import run_vms
 from libs.vm.spec import Interface, Multus, Network
 from libs.vm.vm import BaseVirtualMachine
 from tests.network.localnet.liblocalnet import (
@@ -14,7 +15,6 @@ from tests.network.localnet.liblocalnet import (
     LOCALNET_OVS_BRIDGE_NETWORK,
     LOCALNET_VM_ANTI_AFFINITY,
     localnet_vm,
-    run_vms,
 )
